move exception class to exception pkg
