+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/Ds3Cli.java
@@ -15,6 +15,7 @@
 
 package com.spectralogic.ds3cli;
 
+import com.spectralogic.ds3cli.Exceptions.CommandException;
 import com.spectralogic.ds3cli.command.*;
 import com.spectralogic.ds3cli.util.Ds3Provider;
 import com.spectralogic.ds3cli.util.FileUtils;
