+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/command/GetBucket.java
@@ -17,7 +17,7 @@ package com.spectralogic.ds3cli.command;
 
 
 import com.spectralogic.ds3cli.Arguments;
-import com.spectralogic.ds3cli.CommandException;
+import com.spectralogic.ds3cli.Exceptions.CommandException;
 import com.spectralogic.ds3cli.models.GetBucketResult;
 import com.spectralogic.ds3cli.util.Ds3Provider;
 import com.spectralogic.ds3cli.util.FileUtils;
