+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/views/json/CommandExceptionJsonView.java
@@ -17,7 +17,7 @@ package com.spectralogic.ds3cli.views.json;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.spectralogic.ds3cli.CommandException;
+import com.spectralogic.ds3cli.Exceptions.CommandException;
 import com.spectralogic.ds3cli.View;
 import com.spectralogic.ds3client.networking.FailedRequestException;
